+++ config/config.go
@@ -1,14 +1,41 @@
 package config
 
+import "github.com/gnames/gnfmt"
+
+// Config contains settings necessary for creating index of scientific names
+// of Biodiversity Heritage Library.
 type Config struct {
-	BHLdir      string
-	PgHost      string
-	PgUser      string
-	PgPass      string
-	PgDatabase  string
-	Jobs        int
+	// BHLdir points to a base directory of BHL texts filetree.
+	BHLdir string
+
+	// OutputFormat format of the detected names dump
+	OutputFormat gnfmt.Format
+
+	// PgHost is the IP or a name of a computer running PostgreSQL database.
+	PgHost string
+
+	// PgUser is the username with read/write access to bhlindex database.
+	PgUser string
+
+	// PgPass is the password for PgUser
+	PgPass string
+
+	// PgDatabase is the name of a database for BHLindex.
+	PgDatabase string
+
+	// Jobs is the number of parallel processes running for the name-finding.
+	Jobs int
+
+	// VerifierURL points to a remote GNverifier service.
 	VerifierURL string
+
+	// WithWebLogs can be set to true, if logs from RESTful service are required.
+	// If it set to false, the logs are silenced.
 	WithWebLogs bool
+
+	// WithoutConfirm can be set to true to avoid confirmations before
+	// destructive operations.
+	WithoutConfirm bool
 }
 
 type Option func(*Config)
@@ -16,6 +43,12 @@ type Option func(*Config)
 func OptBHLdir(s string) Option {
 	return func(cfg *Config) {
 		cfg.BHLdir = s
+	}
+}
+
+func OptOutputFormat(f gnfmt.Format) Option {
+	return func(cfg *Config) {
+		cfg.OutputFormat = f
 	}
 }
 
@@ -61,14 +94,21 @@ func OptWithWebLogs(b bool) Option {
 	}
 }
 
+func OptWithoutConfirm(b bool) Option {
+	return func(cfg *Config) {
+		cfg.WithoutConfirm = b
+	}
+}
+
 func New(opts ...Option) Config {
 	res := Config{
-		PgHost:      "localhost",
-		PgUser:      "postgres",
-		PgPass:      "postgres",
-		PgDatabase:  "bhlindex",
-		Jobs:        10,
-		VerifierURL: "https://verifier.globalnames.org/api/v0/",
+		OutputFormat: gnfmt.CSV,
+		PgHost:       "localhost",
+		PgUser:       "postgres",
+		PgPass:       "postgres",
+		PgDatabase:   "bhlindex",
+		Jobs:         10,
+		VerifierURL:  "https://verifier.globalnames.org/api/v0/",
 	}
 	for i := range opts {
 		opts[i](&res)