+++ interface.go
@@ -4,6 +4,7 @@ import (
 	"github.com/gnames/bhlindex/config"
 	"github.com/gnames/bhlindex/ent/finder"
 	"github.com/gnames/bhlindex/ent/loader"
+	"github.com/gnames/bhlindex/ent/output"
 	"github.com/gnames/bhlindex/ent/verif"
 	"github.com/gnames/gnlib/ent/gnvers"
 )
@@ -11,6 +12,7 @@ import (
 type BHLindex interface {
 	FindNames(loader.Loader, finder.Finder) error
 	VerifyNames(verif.VerifierBHL) error
+	DumpNames(output.Dumper) error
 	GetVersion() gnvers.Version
 	GetConfig() config.Config
 }