+++ io/verifio/db-verifio.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"os"
 	"strings"
 	"time"
 
@@ -15,24 +16,24 @@ import (
 
 func (vrf verifio) ExtractUniqueNames() error {
 	log.Info().Msg("Extracting unique name-strings. It will take a while.")
-	q := `INSERT INTO unique_names (name, odds_log10, occurrences, processed)
-          SELECT name, AVG(odds_log10), count(*), false
+	q := `INSERT INTO unique_names (name, odds_log10, occurrences)
+          SELECT name, AVG(odds_log10), count(*)
             FROM detected_names GROUP BY name
             ORDER BY name`
 
 	stmt, err := vrf.db.Prepare(q)
 	if err != nil {
-		return err
+		return fmt.Errorf("ExtractUniqueNames: %w", err)
 	}
 
 	_, err = stmt.Exec()
 	if err != nil {
-		return err
+		return fmt.Errorf("ExtractUniqueNames: %w", err)
 	}
 
 	err = stmt.Close()
 	if err != nil {
-		return err
+		return fmt.Errorf("ExtractUniqueNames: %w", err)
 	}
 	return nil
 }
@@ -40,7 +41,7 @@ func (vrf verifio) ExtractUniqueNames() error {
 func (vrf verifio) checkForDetectedNames() error {
 	noNames, err := vrf.noDetectedNames()
 	if err != nil {
-		return err
+		return fmt.Errorf("checkForDetectedNames: %w", err)
 	}
 	if noNames {
 		err = errors.New("detected_names table is empty")
@@ -70,8 +71,13 @@ func (vrf verifio) truncateVerifTables() error {
 		q := fmt.Sprintf("TRUNCATE TABLE %s RESTART IDENTITY", v)
 		_, err := vrf.db.Exec(q)
 		if err != nil {
-			return err
+			return fmt.Errorf("truncateVerifTables: %w", err)
 		}
+	}
+	q := "ALTER TABLE verified_names DROP CONSTRAINT IF EXISTS verified_names_pkey"
+	_, err := vrf.db.Exec(q)
+	if err != nil {
+		return fmt.Errorf("truncateVerifTables: %w", err)
 	}
 	return nil
 }
@@ -81,49 +87,39 @@ func (vrf verifio) loadNames(
 	namesNum int,
 	chIn chan<- []name.UniqueName,
 ) error {
-	var count int
-	batchSize := 5_000
+	offset := 1
+	limit := 5_000
 
 	q := `
-    WITH temp AS (
-      SELECT id, name, odds_log10, occurrences FROM unique_names
-        WHERE processed=false LIMIT $1)
-    UPDATE unique_names ns SET processed=true
-    FROM temp
-      WHERE ns.name = temp.name
-    RETURNING temp.id, temp.name, temp.odds_log10, temp.occurrences`
-
-	for count < namesNum {
-		rows, err := vrf.db.Query(q, batchSize)
+SELECT id, name, odds_log10, occurrences
+  FROM unique_names
+  WHERE id >= $1 and id < $2
+`
+	for offset < namesNum {
+		rows, err := vrf.db.Query(q, offset, offset+limit)
 		if err != nil {
-			return err
+			return fmt.Errorf("loadNames: %w", err)
 		}
 		var n string
 		var odds float64
 		var id, occur int
-		uns := make([]name.UniqueName, batchSize)
-		var i int
+		uns := make([]name.UniqueName, 0, limit)
 		for rows.Next() {
 			err := rows.Scan(&id, &n, &odds, &occur)
 			if err != nil {
 				rows.Close()
-				return err
+				return fmt.Errorf("loadNames: %w", err)
 			}
-			uns[i] = name.UniqueName{
+			uns = append(uns, name.UniqueName{
 				ID: id, Name: n, OddsLog10: odds, Occurrences: occur,
-			}
-			i++
+			})
 		}
 		rows.Close()
 
-		count = count + batchSize
-		if count > namesNum {
-			uns = uns[0:i]
-		}
+		offset += limit
 		select {
 		case <-ctx.Done():
-			log.Warn().Err(ctx.Err()).Msg("Exiting loading names")
-			return ctx.Err()
+			return fmt.Errorf("loadNames: %w", ctx.Err())
 		case chIn <- uns:
 		}
 	}
@@ -158,11 +154,11 @@ func (vrf verifio) saveVerif(
 	for vns := range chVer {
 		err := vrf.saveNamesToDB(vns)
 		if err != nil {
-			return err
+			return fmt.Errorf("saveVerif: %w", err)
 		}
 		count = incrLog(start, namesNum, count, len(vns))
 	}
-	fmt.Print("\r")
+	fmt.Fprint(os.Stderr, "\r")
 	makeLog(start, namesNum, count)
 	return nil
 }
@@ -170,11 +166,12 @@ func (vrf verifio) saveVerif(
 func incrLog(start time.Time, total, count, incr int) int {
 	count += incr
 	if count%100_000 == 0 {
-		fmt.Print("\r")
+		fmt.Fprint(os.Stderr, "\r")
 		makeLog(start, total, count)
 	} else if count%100 == 0 {
-		fmt.Printf("\r%s", strings.Repeat(" ", 60))
-		fmt.Printf(
+		fmt.Fprintf(os.Stderr, "\r%s", strings.Repeat(" ", 60))
+		fmt.Fprintf(
+			os.Stderr,
 			"\rVerified %s names, %s items/hour",
 			humanize.Comma(int64(count)),
 			namesPerHour(start, count),
@@ -187,43 +184,45 @@ func (vrf verifio) saveNamesToDB(names []name.VerifiedName) error {
 	now := time.Now()
 	columns := []string{
 		"name_id", "name", "record_id", "match_type", "edit_distance",
-		"stem_edit_distance", "matched_name", "matched_canonical", "current_name",
-		"current_canonical", "classification", "data_source_id",
+		"stem_edit_distance", "matched_name", "matched_canonical",
+		"matched_cardinality", "current_name", "current_canonical",
+		"current_cardinality", "classification", "data_source_id",
 		"data_source_title", "data_sources_number", "curation", "odds_log10",
 		"occurrences", "retries", "error", "updated_at",
 	}
 	transaction, err := vrf.db.Begin()
 	if err != nil {
-		return err
+		return fmt.Errorf("saveNamesToDB: %w", err)
 	}
 	defer transaction.Rollback()
 
 	stmt, err := transaction.Prepare(pq.CopyIn("verified_names", columns...))
 	if err != nil {
-		return err
+		return fmt.Errorf("saveNamesToDB: %w", err)
 	}
 
 	for _, v := range names {
 		_, err = stmt.Exec(
 			v.NameID, v.Name, v.RecordID, v.MatchType, v.EditDistance,
-			v.StemEditDistance, v.MatchedName, v.MatchedCanonical, v.CurrentName,
-			v.CurrentCanonical, v.Classification, v.DataSourceID, v.DataSourceTitle,
-			v.DataSourcesNumber, v.Curation, v.OddsLog10, v.Occurrences, v.Retries,
-			v.Error, now,
+			v.StemEditDistance, v.MatchedName, v.MatchedCanonical,
+			v.MatchedCardinality, v.CurrentName, v.CurrentCanonical,
+			v.CurrentCardinality, v.Classification, v.DataSourceID,
+			v.DataSourceTitle, v.DataSourcesNumber, v.Curation, v.OddsLog10,
+			v.Occurrences, v.Retries, v.Error, now,
 		)
 		if err != nil {
-			return err
+			return fmt.Errorf("saveNamesToDB: %w", err)
 		}
 	}
 
 	_, err = stmt.Exec()
 	if err != nil {
-		return err
+		return fmt.Errorf("saveNamesToDB: %w", err)
 	}
 
 	err = stmt.Close()
 	if err != nil {
-		return err
+		return fmt.Errorf("saveNamesToDB: %w", err)
 	}
 
 	return transaction.Commit()